[State] Refactore content of state_tracker.rs
First commit in a row to rearrange the internal structure of the crate.

These commits are necessary to avoid naming confusion when we introduce
the new deserialization trait which will also show up with a new naming
scheme and a second error type.
