+++ src/lib.rs
@@ -6,7 +6,9 @@
 //! The encoder is likewise designed to ensure that it only produces valid structures.
 pub mod decoder;
 pub mod encoder;
+
 mod state_tracker;
+mod token;
 
 use failure::Fail;
